# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GrooveLab API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loops:
    get:
      operationId: getLoops
      tags: [loops]
      summary: List loops with optional filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: timeSignatureId
          in: query
          schema:
            type: integer
        - name: feelId
          in: query
          schema:
            type: integer
        - name: genreId
          in: query
          schema:
            type: integer
        - name: instrumentTypeId
          in: query
          schema:
            type: integer
        - name: contentTypeId
          in: query
          schema:
            type: integer
        - name: bpmMin
          in: query
          schema:
            type: integer
        - name: bpmMax
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [quality_score, newest, most_views, bpm_asc, bpm_desc]
            default: quality_score
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of loops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoopListResponse"
    post:
      operationId: createLoop
      tags: [loops]
      summary: Create a new loop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoopBody"
      responses:
        "201":
          description: Created loop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loop"
  /loops/{id}:
    get:
      operationId: getLoop
      tags: [loops]
      summary: Get a single loop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loop detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loop"
        "404":
          description: Not found
  /loops/{id}/groove-match:
    get:
      operationId: getGrooveMatch
      tags: [loops]
      summary: Get compatible loops (groove match) for a given loop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Compatible loops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loop"
...
  /playlists:
    get:
      operationId: getPlaylists
      tags: [playlists]
      summary: Get user's playlists
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaylistBody"
      responses:
        "201":
          description: Created playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /practice/log:
    post:
      operationId: logPractice
      tags: [practice]
      summary: Log a practice session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeLogBody"
      responses:
        "201":
          description: Logged
  /practice/stats:
    get:
      operationId: getPracticeStats
      tags: [practice]
      summary: Get practice stats for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Practice stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeStats"
...
components:
  schemas:
...
    TimeSignature:
      type: object
      properties:
        id:
          type: integer
        numerator:
          type: integer
        denominator:
          type: integer
        displayName:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - numerator
        - denominator
        - displayName
        - category
        - sortOrder
    Feel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - name
        - sortOrder
    Genre:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentGenreId:
          type: integer
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - name
        - sortOrder
    InstrumentType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - name
        - category
        - sortOrder
    ContentType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - name
        - sortOrder
    Creator:
      type: object
      properties:
        id:
          type: string
        youtubeChannelId:
          type: string
          nullable: true
        channelName:
          type: string
        channelUrl:
          type: string
          nullable: true
        subscriberCount:
          type: integer
        totalViews:
          type: integer
        videoCount:
          type: integer
        description:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        qualityScore:
          type: number
        isVerified:
          type: boolean
        isClaimed:
          type: boolean
      required:
        - id
        - channelName
        - subscriberCount
        - qualityScore
        - isVerified
        - isClaimed
    Loop:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        sourceType:
          type: string
        youtubeVideoId:
          type: string
          nullable: true
        youtubeEmbedUrl:
          type: string
          nullable: true
        midiFileUrl:
          type: string
          nullable: true
        externalUrl:
          type: string
          nullable: true
        creatorId:
          type: string
          nullable: true
        creator:
          $ref: "#/components/schemas/Creator"
          nullable: true
        bpm:
          type: integer
          nullable: true
        bpmRangeLow:
          type: integer
          nullable: true
        bpmRangeHigh:
          type: integer
          nullable: true
        keySignature:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        viewCount:
          type: integer
        likeCount:
          type: integer
        qualityScore:
          type: number
        isEmbeddable:
          type: boolean
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        timeSignatures:
          type: array
          items:
            $ref: "#/components/schemas/TimeSignature"
        feels:
          type: array
          items:
            $ref: "#/components/schemas/Feel"
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        instrumentTypes:
          type: array
          items:
            $ref: "#/components/schemas/InstrumentType"
        contentTypes:
          type: array
          items:
            $ref: "#/components/schemas/ContentType"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - sourceType
        - viewCount
        - likeCount
        - qualityScore
        - isEmbeddable
        - tags
        - isFeatured
        - timeSignatures
        - feels
        - genres
        - instrumentTypes
        - contentTypes
        - createdAt
        - updatedAt
    LoopListResponse:
      type: object
      properties:
        loops:
          type: array
          items:
            $ref: "#/components/schemas/Loop"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - loops
        - total
        - page
        - limit
        - totalPages
...
    CreateLoopBody:
      type: object
      properties:
        title:
          type: string
        sourceType:
          type: string
        youtubeVideoId:
          type: string
        bpm:
          type: integer
        keySignature:
          type: string
        description:
          type: string
        timeSignatureIds:
          type: array
          items:
            type: integer
        feelIds:
          type: array
          items:
            type: integer
        genreIds:
          type: array
          items:
            type: integer
        instrumentTypeIds:
          type: array
          items:
            type: integer
        contentTypeIds:
          type: array
          items:
            type: integer
      required:
        - title
        - sourceType
...
    Playlist:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        isPublic:
          type: boolean
        loops:
          type: array
          items:
            $ref: "#/components/schemas/Loop"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - isPublic
        - loops
        - createdAt
    CreatePlaylistBody:
      type: object
      properties:
        userId:
          type: string
        title:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
      required:
        - userId
        - title
    PracticeLogBody:
      type: object
      properties:
        userId:
          type: string
        loopId:
          type: string
        durationSeconds:
          type: integer
        bpmPracticed:
          type: integer
      required:
        - userId
        - durationSeconds
    PracticeStats:
      type: object
      properties:
        totalSessions:
          type: integer
        totalMinutes:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        recentLogs:
          type: array
          items:
            $ref: "#/components/schemas/PracticeLog"
      required:
        - totalSessions
        - totalMinutes
        - currentStreak
        - longestStreak
        - recentLogs
    PracticeLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        loopId:
          type: string
          nullable: true
        durationSeconds:
          type: integer
        bpmPracticed:
          type: integer
          nullable: true
        practicedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - durationSeconds
        - practicedAt
...