# lib/api-zod/src/generated/types/createLoopBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrooveLab API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLoopBody {
  title: string;
  sourceType: string;
  youtubeVideoId?: string;
  bpm?: number;
  keySignature?: string;
  description?: string;
  timeSignatureIds?: number[];
  feelIds?: number[];
  genreIds?: number[];
  instrumentTypeIds?: number[];
  contentTypeIds?: number[];
}