# lib/api-zod/src/generated/types/loopListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrooveLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Loop } from "./loop";

export interface LoopListResponse {
  loops: Loop[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}