# lib/api-zod/src/generated/types/practiceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GrooveLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeLog } from "./practiceLog";

export interface PracticeStats {
  totalSessions: number;
  totalMinutes: number;
  currentStreak: number;
  longestStreak: number;
  recentLogs: PracticeLog[];
}